Deduplicate modal test fixtures into imbue.mng_modal.conftest
Move shared modal test fixtures (setup_test_mng_env with token loading,
modal_subprocess_env and its session dependencies) into the inner conftest
at imbue.mng_modal.conftest so they are importable by other packages.

mng_claude now imports these via pytest_plugins instead of duplicating
~90 lines of identical fixture code. Added mng-modal as a dependency
of mng_claude to support this import.
