+++ libs/mng_claude/pyproject.toml
@@ -6,6 +6,7 @@ readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
     "mng==0.1.7",
+    "mng-modal==0.1.0",
 ]
 
 [project.entry-points.mng]
@@ -22,6 +23,7 @@ packages = ["imbue"]
 
 [tool.uv.sources]
 mng = { workspace = true }
+mng-modal = { workspace = true }
 
 # Shared pytest settings (markers, filterwarnings, coverage report config)
 # are centralized in:
